remove impl/serializer.hpp
